refactor: update receipt service imports and enhance test consistency
- Changed import paths for receipt services to improve code organization across multiple test files.
- Updated draft ID type from number to string in receipt-related tests to align with API responses.
- Adjusted receipt creation and deletion logic in tests to reflect new API structure and ensure accurate data handling.
